backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
supabase
python-multipart
pytesseract
pillow
python-dotenv
aiohttp
